Extract types to types/field-api
